Splitting pull request data
